retoques exposiciones endpoints

+++ Agregador/src/main/java/ar/utn/dssi/Agregador/models/DTOs/inputDTO/FiltroInputDTO.java
@@ -14,5 +14,5 @@ public class FiltroInputDTO {
   private LocalDate fecha_acontecimiento_hasta;
   private Double longitud;
   private Double latitud;
-  private Integer idFuente;
+  private Long idFuente;
 }
